fix: satisfy advisor source size ratchet
